refactor: 변경 Detail 컴포넌트
* Detail 컴포넌트 DetailHeader 컴포넌트로 분리
